+++ SimulationCode/src/StepNaive.cpp
@@ -1,6 +1,6 @@
 #include "StepNaive.h"
 
-StepNaive::StepNaive(std::vector<float> &masses, unsigned numberOfBodies) {
+StepNaive::StepNaive(std::vector<double> &masses, unsigned numberOfBodies) {
   firstStep = false;
   this->numberOfBodies = numberOfBodies;
   forces.resize(3 * numberOfBodies);
@@ -8,7 +8,7 @@ StepNaive::StepNaive(std::vector<float> &masses, unsigned numberOfBodies) {
   for (unsigned i = 0; i < numberOfBodies; i++)
     weights[i] = masses[i];
   randomGenerator = new RandomGenerators();
-  randomGenerator->initializeVelocities<std::vector<float>>(velocities, numberOfBodies);
+  randomGenerator->initializeVelocities<std::vector<double>>(velocities, numberOfBodies);
 }
 
 StepNaive::~StepNaive() {
@@ -19,7 +19,7 @@ StepNaive::~StepNaive() {
 }
 
 bool StepNaive::testingMomemntum() {
-    float momentumX = 0.0f, momentumY = 0.0f, momentumZ = 0.0f;
+    double momentumX = 0.0f, momentumY = 0.0f, momentumZ = 0.0f;
     for (unsigned i = 0; i < numberOfBodies; i++) {
         momentumX += (weights[i] * velocities[i * 3]);
         momentumY += (weights[i] * velocities[i * 3 + 1]);
@@ -43,20 +43,20 @@ bool StepNaive::testingMomemntum() {
     return true;
 }
 
-void StepNaive::compute(tf3 &positions, float dt) {
+void StepNaive::compute(tf3 &positions, double dt) {
   testingMomemntum();
   //assert(testingMomemntum());
-  float EPS = 0.01;
+  double EPS = 0.01;
   std::fill(forces.begin(), forces.end(), 0.0);
   for (unsigned i = 0; i < numberOfBodies; i++) {
     for (unsigned j = 0; j < numberOfBodies; j++) {
-      float distX = positions[j * 3] - positions[i * 3];
-      float distY = positions[j * 3 + 1] - positions[i * 3 + 1];
-      float distZ = positions[j * 3 + 2] - positions[i * 3 + 2];
-      float dist = (distX * distX + distY * distY + distZ * distZ) + EPS * EPS;
+      double distX = positions[j * 3] - positions[i * 3];
+      double distY = positions[j * 3 + 1] - positions[i * 3 + 1];
+      double distZ = positions[j * 3 + 2] - positions[i * 3 + 2];
+      double dist = (distX * distX + distY * distY + distZ * distZ) + EPS * EPS;
       dist = dist * sqrt(dist);
       if (i != j) {
-        float F = G * (weights[i] * weights[j]);
+        double F = G * (weights[i] * weights[j]);
         forces[i * 3] += F * distX / dist; // force = G(m1*m2)/r^2
         forces[i * 3 + 1] += F * distY / dist;
         forces[i * 3 + 2] += F * distZ / dist;
@@ -65,7 +65,7 @@ void StepNaive::compute(tf3 &positions, float dt) {
   }
   for (unsigned i = 0; i < numberOfBodies; i++) {
     for (int j = 0; j < 3; j++) {
-      float acceleration = forces[i * 3 + j] / weights[i];
+      double acceleration = forces[i * 3 + j] / weights[i];
       positions[i * 3 + j] +=
           velocities[i * 3 + j] * dt + acceleration * dt * dt / 2;
       velocities[i * 3 + j] += acceleration * dt;
